+++ core/src/main/java/tech/ydb/core/impl/YdbSchedulerFactory.java
@@ -1,9 +1,17 @@
 package tech.ydb.core.impl;
 
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
+import java.util.concurrent.Callable;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import javax.annotation.Nonnull;
@@ -53,6 +61,10 @@ public class YdbSchedulerFactory {
         }
     }
 
+    public static ScheduledExecutorService wrapExternal(ScheduledExecutorService scheduler) {
+        return new NonStoppableScheduler(scheduler);
+    }
+
     private static class YdbThreadFactory implements ThreadFactory {
         private static final AtomicInteger INSTANCE_COUNT = new AtomicInteger(1);
 
@@ -70,6 +82,104 @@ public class YdbSchedulerFactory {
             Thread t = new Thread(r, namePrefix + threadNumber.getAndIncrement());
             t.setDaemon(true);
             return t;
+        }
+    }
+
+
+    private static class NonStoppableScheduler implements ScheduledExecutorService {
+        private final ScheduledExecutorService origin;
+
+        NonStoppableScheduler(ScheduledExecutorService service) {
+            this.origin = service;
+        }
+
+        @Override
+        public void shutdown() {
+            // Nothing
+        }
+
+        @Override
+        public List<Runnable> shutdownNow() {
+            return Collections.emptyList();
+        }
+
+        @Override
+        public boolean isShutdown() {
+            return false;
+        }
+
+        @Override
+        public boolean isTerminated() {
+            return false;
+        }
+
+        @Override
+        public boolean awaitTermination(long timeout, TimeUnit unit) throws InterruptedException {
+            return true;
+        }
+
+        @Override
+        public ScheduledFuture<?> schedule(Runnable command, long delay, TimeUnit unit) {
+            return origin.schedule(command, delay, unit);
+        }
+
+        @Override
+        public <V> ScheduledFuture<V> schedule(Callable<V> callable, long delay, TimeUnit unit) {
+            return origin.schedule(callable, delay, unit);
+        }
+
+        @Override
+        public ScheduledFuture<?> scheduleAtFixedRate(Runnable command, long initialDelay, long period, TimeUnit unit) {
+            return origin.scheduleAtFixedRate(command, initialDelay, period, unit);
+        }
+
+        @Override
+        public ScheduledFuture<?> scheduleWithFixedDelay(Runnable command, long initialDelay, long delay,
+                TimeUnit unit) {
+            return origin.scheduleWithFixedDelay(command, initialDelay, delay, unit);
+        }
+
+        @Override
+        public <T> Future<T> submit(Callable<T> task) {
+            return origin.submit(task);
+        }
+
+        @Override
+        public <T> Future<T> submit(Runnable task, T result) {
+            return origin.submit(task, result);
+        }
+
+        @Override
+        public Future<?> submit(Runnable task) {
+            return origin.submit(task);
+        }
+
+        @Override
+        public <T> List<Future<T>> invokeAll(Collection<? extends Callable<T>> tasks) throws InterruptedException {
+            return origin.invokeAll(tasks);
+        }
+
+        @Override
+        public <T> List<Future<T>> invokeAll(Collection<? extends Callable<T>> tasks, long timeout, TimeUnit unit)
+                throws InterruptedException {
+            return origin.invokeAll(tasks, timeout, unit);
+        }
+
+        @Override
+        public <T> T invokeAny(Collection<? extends Callable<T>> tasks) throws InterruptedException,
+                ExecutionException {
+            return origin.invokeAny(tasks);
+        }
+
+        @Override
+        public <T> T invokeAny(Collection<? extends Callable<T>> tasks, long timeout, TimeUnit unit)
+                throws InterruptedException, ExecutionException, TimeoutException {
+            return origin.invokeAny(tasks, timeout, unit);
+        }
+
+        @Override
+        public void execute(Runnable command) {
+            origin.execute(command);
         }
     }
 }