+++ core/src/main/java/tech/ydb/core/impl/pool/ShadedNettyChannelFactory.java
@@ -10,7 +10,6 @@ import javax.net.ssl.SSLException;
 import io.grpc.ClientInterceptor;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
-import io.grpc.Metadata;
 import io.grpc.internal.DnsNameResolverProvider;
 import io.grpc.netty.shaded.io.grpc.netty.GrpcSslContexts;
 import io.grpc.netty.shaded.io.grpc.netty.NegotiationType;
@@ -19,7 +18,6 @@ import io.grpc.netty.shaded.io.netty.buffer.ByteBufAllocator;
 import io.grpc.netty.shaded.io.netty.channel.ChannelOption;
 import io.grpc.netty.shaded.io.netty.handler.ssl.SslContext;
 import io.grpc.netty.shaded.io.netty.handler.ssl.SslContextBuilder;
-import io.grpc.stub.MetadataUtils;
 
 import tech.ydb.core.grpc.GrpcTransportBuilder;
 import tech.ydb.core.grpc.YdbHeaders;
@@ -33,8 +31,8 @@ public class ShadedNettyChannelFactory implements ManagedChannelFactory {
     static final int INBOUND_MESSAGE_SIZE = 64 << 20; // 64 MiB
     static final String DEFAULT_BALANCER_POLICY = "round_robin";
 
-    private final String database;
-    private final String version;
+    private final ClientInterceptor metadata;
+
     private final boolean useTLS;
     private final byte[] cert;
     private final boolean retryEnabled;
@@ -42,9 +40,8 @@ public class ShadedNettyChannelFactory implements ManagedChannelFactory {
     private final Long grpcKeepAliveTimeMillis;
     private final List<Consumer<? super ManagedChannelBuilder<?>>> initializers;
 
-    public ShadedNettyChannelFactory(GrpcTransportBuilder builder) {
-        this.database = builder.getDatabase();
-        this.version = builder.getVersionString();
+    private ShadedNettyChannelFactory(GrpcTransportBuilder builder) {
+        this.metadata = YdbHeaders.createMetadataInterceptor(builder);
         this.useTLS = builder.getUseTls();
         this.cert = builder.getCert();
         this.retryEnabled = builder.isEnableRetry();
@@ -74,7 +71,7 @@ public class ShadedNettyChannelFactory implements ManagedChannelFactory {
                 .maxInboundMessageSize(INBOUND_MESSAGE_SIZE)
                 .withOption(ChannelOption.ALLOCATOR, ByteBufAllocator.DEFAULT)
                 .withOption(ChannelOption.TCP_NODELAY, true)
-                .intercept(metadataInterceptor());
+                .intercept(metadata);
 
         if (!useDefaultGrpcResolver) {
             // force usage of dns resolver and round_robin balancer
@@ -105,13 +102,6 @@ public class ShadedNettyChannelFactory implements ManagedChannelFactory {
 
     protected void configure(NettyChannelBuilder channelBuilder) {
 
-    }
-
-    private ClientInterceptor metadataInterceptor() {
-        Metadata extraHeaders = new Metadata();
-        extraHeaders.put(YdbHeaders.DATABASE, database);
-        extraHeaders.put(YdbHeaders.BUILD_INFO, version);
-        return MetadataUtils.newAttachHeadersInterceptor(extraHeaders);
     }
 
     private SslContext createSslContext() {